requirements.txt
# Core
flask==3.1.0
feedparser==6.0.11
openpyxl==3.1.5
requests==2.32.3

# Resume parsing
pypdf>=4.0.0
python-docx>=1.0.0
docx2pdf>=0.1.8

# AI — Claude API (primary)
anthropic>=0.34.0

# Browser automation (CV PDF generation + form filling)
playwright>=1.40.0

# Config
pyyaml>=6.0.1

# Optional: Ollama fallback (not required)
# ollama>=0.1.0
